+++ golem/core/__init__.py
@@ -1,14 +1,14 @@
-import sys
-import os
-import types
+#import sys
+#import os
+#import types
 
 #from selenium import webdriver
 #from selenium.webdriver.common.desired_capabilities import DesiredCapabilities
 #from selenium.common.exceptions import WebDriverException
 
-from golem.core import execution_logger as logger
-from golem.selenium import _find, _find_all
-from golem import execution
+#from golem.core import execution_logger as logger
+#from golem.selenium import _find, _find_all
+#from golem import execution
 
 
 # def get_or_create_webdriver():