+++ CHANGELOG.md
@@ -6,8 +6,12 @@
 ### Added
 - Documented the Golem Public API
 - UP, DOWN, LEFT, RIGHT options to actions.press_key (thanks to danielmaddern)
+- Add counter of total and selected tests in suite view
+- golem.execution module to track data for current execution
 
-###
+### Changed
+- golem.selenium is now golem.browser
+- golem.selenium.get_driver is now golem.browser.get_browser
 
 
 ## [0.1.22] - 2017-09-27